[Ref] Replaced some CArrays with vectors [Mod] Pattern Editor: Interpolation and transpose context menu items have been moved to sub menus.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@1295 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/AutoSaver.h
@@ -56,7 +56,7 @@ private:
 	bool m_bEnabled;
 	DWORD m_nLastSave;
 	DWORD m_nSaveInterval;
-	int m_nBackupHistory;
+	size_t m_nBackupHistory;
 	bool m_bUseOriginalPath;
 	CString m_csPath;
 	CString m_csFileNameTemplate;
@@ -94,7 +94,7 @@ public:
 	virtual BOOL OnInitDialog();
 	afx_msg void OnBnClickedAutosaveEnable();
 	afx_msg void OnBnClickedAutosaveUseorigdir();
-    void OnSettingsChanged();
+	void OnSettingsChanged();
 	BOOL OnSetActive();
 	
 };
